feat: Move all files for packing into `out` subdirectory (#3)
This change to the pre packer works off a list of the output files we would like to pack, as well as the original input file (placed in a data file within the working directory either by the court document parse step function or at the bag unpacking stage in the case of requests from TDR.

## Test Plan

Pushed updated version to ECR and updated pte to use it. Used system testing repo to test end to end for FCL and verified pack appearance as expected.